% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thank.R
\name{to}
\alias{to}
\title{Send an email}
\usage{
to(package)
}
\arguments{
\item{package, }{name of the package you want to say thanks to.}
}
\description{
Opens a prepopulated email for you to send to the maintainer of a package.
}
